+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+# UNRELEASED
+
+- Update metrics collection to remove labels [#163](https://github.com/gohypermode/runtime/pull/163)
+- Add environment and version to health endpoint [#164](https://github.com/gohypermode/runtime/pull/164)
+
 # 2024-04-29 - Version 0.6.2
 
 - Traces and non-user errors are now sent to Sentry [#158](https://github.com/gohypermode/runtime/issues/158)